backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
sentence-transformers
pydantic
qdrant-client
numpy